refactor: simplify test context creation for ProductsService and ShoppingLists
